scripto/flask/web_with_forms/requirements.txt
Flask==1.1.2
google-cloud-storage
pandas
json2html
flask-wtf
